requirements.txt
calmjs.parse
crispy-bootstrap5
csscompressor
Django==5.2
django-cleanup
django_compressor
django-crispy-forms
django-debug-toolbar
django-environ
django-extensions
django-filter
django-fullurl
django-pylibmc
django-rest-framework
flake8
gunicorn
html5lib
markdown
Pillow
python-dotenv
psycopg2-binary
requests
surlex
sorl-thumbnail
whitenoise